added support for double

+++ include/read_file.h
@@ -2,6 +2,6 @@
 #define READ_FILE_H
 
 std::unordered_map<std::string, std::vector<std::string>> read_metadata(std::string);
-std::vector<long long int> read_table_column(std::string, int);
+std::vector<double> read_table_column(std::string, int);
 
 #endif
